feat: renamed components files for more consistence and updated the views that use this components.
